refactor(gpu-ir): centralize shared attribute types
